cmake

+++ perf_tests/xe_peak/src/options.cpp
@@ -1,24 +1,24 @@
 /*
- * Copyright(c) 2019 Intel Corporation
+ * INTEL CONFIDENTIAL
+ * Copyright (c) 2016 - 2019 Intel Corporation. All Rights Reserved.
  *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
+ * The source code contained or described herein and all documents related to
+ * the source code ("Material") are owned by Intel Corporation or its suppliers
+ * or licensors. Title to the Material remains with Intel Corporation or its
+ * suppliers and licensors. The Material contains trade secrets and proprietary
+ * and confidential information of Intel or its suppliers and licensors. The
+ * Material is protected by worldwide copyright and trade secret laws and
+ * treaty provisions. No part of the Material may be used, copied, reproduced,
+ * modified, published, uploaded, posted, transmitted, distributed, or
+ * disclosed in any way without Intel's prior express written permission.
  *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
+ * No license under any patent, copyright, trade secret or other intellectual
+ * property right is granted to or conferred upon you by disclosure or delivery
+ * of the Materials, either expressly, by implication, inducement, estoppel or
+ * otherwise. Any license under such intellectual property rights must be
+ * express and approved by Intel in writing.
  */
+
 #include "../include/xe_peak.h"
 
 static const char *usage_str =
@@ -63,13 +63,13 @@ int XePeak::parse_arguments(int argc, char **argv) {
     } else if ((strcmp(argv[i], "-p") == 0) ||
                (strcmp(argv[i], "--platform") == 0)) {
       if ((i + 1) < argc) {
-        specified_platform = strtoul(argv[i + 1], NULL, 0);
+        specified_platform = static_cast<int>(strtoul(argv[i + 1], NULL, 0));
         i++;
       }
     } else if ((strcmp(argv[i], "-d") == 0) ||
                (strcmp(argv[i], "--device") == 0)) {
       if ((i + 1) < argc) {
-        specified_device = strtoul(argv[i + 1], NULL, 0);
+        specified_device = static_cast<int>(strtoul(argv[i + 1], NULL, 0));
         i++;
       }
     } else if (strcmp(argv[i], "-e") == 0) {
@@ -78,12 +78,12 @@ int XePeak::parse_arguments(int argc, char **argv) {
       verbose = true;
     } else if (strcmp(argv[i], "-i") == 0) {
       if ((i + 1) < argc) {
-        iters = strtoul(argv[i + 1], NULL, 0);
+        iters = static_cast<uint32_t>(strtoul(argv[i + 1], NULL, 0));
         i++;
       }
     } else if (strcmp(argv[i], "-w") == 0) {
       if ((i + 1) < argc) {
-        warmup_iterations = strtoul(argv[i + 1], NULL, 0);
+        warmup_iterations = static_cast<int>(strtoul(argv[i + 1], NULL, 0));
         i++;
       }
     } else if ((strcmp(argv[i], "-t") == 0)) {
